requirements-test.txt
# Testing dependencies
pytest==7.4.3
pytest-flask==1.3.0
pytest-html==4.1.1
pytest-cov==4.1.0
pytest-xdist==3.5.0
requests==2.31.0
factory-boy==3.3.0
faker==20.1.0
